Create README.md
I made this using ren'py. Created by caqw1ii(also known as leizmira).

Music and characters belong to Nankidai, who made "Your turn to die". Script, story, and art are made by caqw1ii.

This is a fan game and does not fully relate to the events in the og(your turn to die) game. The events are happening before the death game starts. In this visual novel, Sara Chidouin and Ranmaru Kageyama are studying at the same school. I tried to link it to the actual events, but mostly this whole story is a work of fiction and something that I came up with. This is also the very first visual novel I've made. I already had drawings that kinda looked like sprites, so I decided to make a yttd fan game out of it. I'm not fluent in English, that's why it might have some minor grammar mistakes.